+++ client/src/App.tsx
@@ -1,25 +1,24 @@
 import { ChakraProvider } from '@chakra-ui/react';
-import React from 'react';
+import React, { useContext } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import QuizTestPage from './pages/quizTestPage';
 import LeftMenuPage from './pages/leftmenuTestPage';
 import ChatTestPage from './pages/chatTestPage';
 import YoutubeTestPage from './pages/youtubeTestPage_combined';
 import LobbyPage from './pages/lobbyPage';
-import LoginPage from './pages/loginPage';
+import LoginModal from './components/loginModal';
+import UserContext from './context/user/userContext';
 
 const App = (): React.ReactElement<any, any> => {
   return (
     <>
       <div className="App">
-        <ChakraProvider>
-          <Router>
-            <Routes>
-              <Route path="/" element={<LobbyPage />} />
-              <Route path="/login" element={<LoginPage />} />
-            </Routes>
-          </Router>
-        </ChakraProvider>
+        <LoginModal />
+        <Router>
+          <Routes>
+            <Route path="/" element={<LobbyPage />} />
+          </Routes>
+        </Router>
       </div>
     </>
   );
