+++ client/src/index.tsx
@@ -1,13 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { ChakraProvider } from '@chakra-ui/react';
 import './index.css';
 import App from './App';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 import reportWebVitals from './reportWebVitals';
+import UserProvider from './context/user/userProvider';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <ChakraProvider>
+      <UserProvider>
+        <App />
+      </UserProvider>
+    </ChakraProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
